Issue #54 - Refactor username display in post profile
Introduced `postUsername` variable to handle the display of the post author's username, defaulting to "Guest" if the user is null. Replaced the `<dd>` element containing the username link with a new `<a>` tag inside the `<dt>` element. The new link includes a colored `<span>` based on the user's color property, defaulting to black if unspecified. These changes improve the structure, styling, and functionality of the username display.